refactor(storage): extract _HealthMixin
Moves 7 health observability methods (fetch_stock_history_rollup,
count_active_watchlist, _discover_record_health_rules, list_record_health,
upsert_heartbeat, get_heartbeat, get_latest_heartbeat) plus 3 module
constants (_RECORD_HEALTH_TIMESTAMP_COLUMNS, _RECORD_HEALTH_TICKER_COLUMNS,
_RECORD_HEALTH_EXCLUDED_TABLES) to storage/health.py.

Constants are used only by _discover_record_health_rules (verified via
grep) so they move with the mixin.
